[HHQ-1078] Long ID unnecessarily downcast to Integer, may cause problems with foreign key, which stores the original Long value [merge from HEAD]

+++ src/org/hyperic/hq/events/server/session/EventTrackerEJBImpl.java
@@ -114,7 +114,7 @@ public class EventTrackerEJBImpl extends SessionBase implements SessionBean {
 
                     // Insert into Event table first
                     int i = 1;
-                    stmt.setInt(i++, event.getId().intValue());
+                    stmt.setLong(i++, event.getId().longValue());
                     stmt.setBytes(i++, ostream.toByteArray());
                     stmt.setLong(i++, event.getTimestamp());
                     stmt.execute();
